requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
PyPDF2>=3.0.0
python-docx>=1.1.0
openpyxl>=3.1.0
python-pptx>=0.6.23
olefile>=0.47
